Update SwitchDialog.java
If this is a scene, ExtendedStatusInformation is not available, and the getStatus call will fail. This renders scenes with a number where not a switch with that same number is available unselectable.
This uncovers the next problem. Scenes can be protected as well, so we need to check that in another way.