Update SwitchDialog.java
The getStatus call was used to get protection status of switch/scene.
However, "info" already has that information in it's structure.
So the whole getStatus call is not needed anymore (as far as I can see). So I removed it, and simplified code a little.
Tested with protected/nonprotected switch and scene. Works fine in my test environment.